# Move statics to a separate class and add JsonWriter perf test.